+++ src/main/kotlin/com/bridgecrew/services/checkovService/PipenvCheckovService.kt
@@ -1,77 +1,77 @@
-package com.bridgecrew.services.checkovService
-
-import CliService
-import com.intellij.openapi.components.service
-import com.intellij.openapi.diagnostic.logger
-import com.intellij.openapi.project.Project
-import org.apache.commons.io.FilenameUtils
-
-private val LOG = logger<PipenvCheckovService>()
-
-class PipenvCheckovService(project: Project) : CheckovService(project) {
-
-    override fun getInstallCommand(): ArrayList<String> {
-        val cmds = arrayListOf("pipenv","--python","3","install","checkov")
-        return cmds
-    }
-
-//    override fun getExecCommand(filePath: String, apiToken: String, gitRepoName: String, prismaUrl: String?): ArrayList<String> {
-//        val relevantFilePath = FilenameUtils.separatorsToSystem(filePath)
-//        val cmds = arrayListOf(project.service<CliService>().checkovPath, "-s","--bc-api-key",
-//            apiToken, "--repo-id", gitRepoName, "-f", relevantFilePath,"-o", "json")
+//package com.bridgecrew.services.checkovService
+//
+//import CliService
+//import com.intellij.openapi.components.service
+//import com.intellij.openapi.diagnostic.logger
+//import com.intellij.openapi.project.Project
+//import org.apache.commons.io.FilenameUtils
+//
+//private val LOG = logger<PipenvCheckovService>()
+//
+//class PipenvCheckovService(project: Project) : PythonCheckovService(project) {
+//
+//    override fun getInstallCommand(): ArrayList<String> {
+//        val cmds = arrayListOf("pipenv","--python","3","install","checkov")
 //        return cmds
 //    }
-
-    override fun getVersion(project: Project): ArrayList<String> {
-        val cmds = arrayListOf(project.service<CliService>().checkovPath,"-v")
-        return cmds
-    }
-
-    override fun getCheckovRunningCommandByServiceType(): ArrayList<String> {
-        return arrayListOf(project.service<CliService>().checkovPath)
-    }
-
-    override fun getDirectory(): String {
-        return FilenameUtils.separatorsToSystem(project.basePath!!)
-    }
-
-    companion object {
-        fun getCheckovPath(project: Project) {
-            val os = System.getProperty("os.name").toLowerCase()
-            if (os.contains("win")) {
-                val command = arrayListOf("pipenv", "run", "where", "python")
-                project.service<CliService>().run(command, project, Companion::updateCheckovPathWin)
-            } else {
-                val command = arrayListOf("pipenv", "run", "which", "python")
-                project.service<CliService>().run(command, project, Companion::updateCheckovPathUnix)
-
-            }
-        }
-
-
-        fun updateCheckovPathUnix(output: String, exitCode: Int, project: Project) {
-            if (exitCode != 0 || output.contains("[ERROR]")) {
-                LOG.warn("Failed to get checkovPath")
-                return
-            }
-            val result = output.trim()
-            val checkovPathArray: MutableList<String> = result.split('/').toMutableList()
-            checkovPathArray.removeLast()
-            checkovPathArray.add("checkov")
-            project.service<CliService>().checkovPath = checkovPathArray.joinToString(separator = "/")
-            LOG.info("Setting checkovPath: ${project.service<CliService>().checkovPath}")
-        }
-
-        fun updateCheckovPathWin(output: String, exitCode: Int, project: Project) {
-            if (exitCode != 0 || output.contains("[ERROR]")) {
-                LOG.warn("Failed to get checkovPath")
-                return
-            }
-            val result = output.trim()
-            val checkovPathArray = result.split('\n')
-            LOG.info("Checkov path in Win is $result")
-            project.service<CliService>().checkovPath = checkovPathArray[0]
-            LOG.info("Setting checkovPath: ${project.service<CliService>().checkovPath}")
-        }
-    }
-}
+//
+////    override fun getExecCommand(filePath: String, apiToken: String, gitRepoName: String, prismaUrl: String?): ArrayList<String> {
+////        val relevantFilePath = FilenameUtils.separatorsToSystem(filePath)
+////        val cmds = arrayListOf(project.service<CliService>().checkovPath, "-s","--bc-api-key",
+////            apiToken, "--repo-id", gitRepoName, "-f", relevantFilePath,"-o", "json")
+////        return cmds
+////    }
+//
+//    override fun getVersion(project: Project): ArrayList<String> {
+//        val cmds = arrayListOf(project.service<CliService>().checkovPath,"-v")
+//        return cmds
+//    }
+//
+//    override fun getCheckovRunningCommandByServiceType(): ArrayList<String> {
+//        return arrayListOf(project.service<CliService>().checkovPath)
+//    }
+//
+//    override fun getDirectory(): String {
+//        return FilenameUtils.separatorsToSystem(project.basePath!!)
+//    }
+//
+////    companion object {
+////        fun getCheckovPath(project: Project) {
+////            val os = System.getProperty("os.name").toLowerCase()
+////            if (os.contains("win")) {
+////                val command = arrayListOf("pipenv", "run", "where", "python")
+////                project.service<CliService>().run(command, project, Companion::updateCheckovPathWin)
+////            } else {
+////                val command = arrayListOf("pipenv", "run", "which", "python")
+////                project.service<CliService>().run(command, project, Companion::updateCheckovPathUnix)
+////
+////            }
+////        }
+////
+////
+////        fun updateCheckovPathUnix(output: String, exitCode: Int, project: Project) {
+////            if (exitCode != 0 || output.contains("[ERROR]")) {
+////                LOG.warn("Failed to get checkovPath")
+////                return
+////            }
+////            val result = output.trim()
+////            val checkovPathArray: MutableList<String> = result.split('/').toMutableList()
+////            checkovPathArray.removeLast()
+////            checkovPathArray.add("checkov")
+////            project.service<CliService>().checkovPath = checkovPathArray.joinToString(separator = "/")
+////            LOG.info("Setting checkovPath: ${project.service<CliService>().checkovPath}")
+////        }
+////
+////        fun updateCheckovPathWin(output: String, exitCode: Int, project: Project) {
+////            if (exitCode != 0 || output.contains("[ERROR]")) {
+////                LOG.warn("Failed to get checkovPath")
+////                return
+////            }
+////            val result = output.trim()
+////            val checkovPathArray = result.split('\n')
+////            LOG.info("Checkov path in Win is $result")
+////            project.service<CliService>().checkovPath = checkovPathArray[0]
+////            LOG.info("Setting checkovPath: ${project.service<CliService>().checkovPath}")
+////        }
+////    }
+//}
